# build/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"_36_Enums.UserRole": {
				"type": "string",
				"enum": [
					"ADMIN",
					"STAFF",
					"TEACHER",
					"STUDENT"
				]
			},
			"UserRole": {
				"$ref": "#/components/schemas/_36_Enums.UserRole"
			},
			"CreateUserDTO": {
				"properties": {
					"username": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"role": {
						"$ref": "#/components/schemas/UserRole"
					}
				},
				"required": [
					"username",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateUserDTO": {
				"properties": {
					"username": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"role": {
						"$ref": "#/components/schemas/UserRole"
					},
					"isActive": {
						"type": "boolean"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"DefaultSelection_Prisma._36_DepartmentPayload_": {
				"properties": {
					"code": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"code",
					"updatedAt",
					"createdAt",
					"id",
					"name"
				],
				"type": "object"
			},
			"Department": {
				"$ref": "#/components/schemas/DefaultSelection_Prisma._36_DepartmentPayload_",
				"description": "Model Department"
			},
			"DepartmentCreateRequest": {
				"properties": {
					"name": {
						"type": "string"
					},
					"code": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"code"
				],
				"type": "object",
				"additionalProperties": false
			},
			"DepartmentUpdateRequest": {
				"properties": {
					"name": {
						"type": "string"
					},
					"code": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "truongtrandainghia_be",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/users": {
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"updatedAt": {
											"type": "string",
											"format": "date-time"
										},
										"createdAt": {
											"type": "string",
											"format": "date-time"
										},
										"isActive": {
											"type": "boolean"
										},
										"role": {
											"$ref": "#/components/schemas/_36_Enums.UserRole"
										},
										"email": {
											"type": "string"
										},
										"password": {
											"type": "string"
										},
										"username": {
											"type": "string"
										},
										"id": {
											"type": "number",
											"format": "double"
										}
									},
									"required": [
										"updatedAt",
										"createdAt",
										"isActive",
										"role",
										"email",
										"password",
										"username",
										"id"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateUserDTO"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "FindAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"properties": {
											"updatedAt": {
												"type": "string",
												"format": "date-time"
											},
											"createdAt": {
												"type": "string",
												"format": "date-time"
											},
											"isActive": {
												"type": "boolean"
											},
											"role": {
												"$ref": "#/components/schemas/_36_Enums.UserRole"
											},
											"email": {
												"type": "string"
											},
											"password": {
												"type": "string"
											},
											"username": {
												"type": "string"
											},
											"id": {
												"type": "number",
												"format": "double"
											}
										},
										"required": [
											"updatedAt",
											"createdAt",
											"isActive",
											"role",
											"email",
											"password",
											"username",
											"id"
										],
										"type": "object"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": []
			}
		},
		"/users/search": {
			"get": {
				"operationId": "QueryUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"properties": {
											"updatedAt": {
												"type": "string",
												"format": "date-time"
											},
											"createdAt": {
												"type": "string",
												"format": "date-time"
											},
											"isActive": {
												"type": "boolean"
											},
											"role": {
												"$ref": "#/components/schemas/_36_Enums.UserRole"
											},
											"email": {
												"type": "string"
											},
											"password": {
												"type": "string"
											},
											"username": {
												"type": "string"
											},
											"id": {
												"type": "number",
												"format": "double"
											}
										},
										"required": [
											"updatedAt",
											"createdAt",
											"isActive",
											"role",
											"email",
											"password",
											"username",
											"id"
										],
										"type": "object"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "email",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "username",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "role",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/UserRole"
						}
					},
					{
						"in": "query",
						"name": "isActive",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/users/{id}": {
			"get": {
				"operationId": "FindById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"updatedAt": {
											"type": "string",
											"format": "date-time"
										},
										"createdAt": {
											"type": "string",
											"format": "date-time"
										},
										"isActive": {
											"type": "boolean"
										},
										"role": {
											"$ref": "#/components/schemas/_36_Enums.UserRole"
										},
										"email": {
											"type": "string"
										},
										"password": {
											"type": "string"
										},
										"username": {
											"type": "string"
										},
										"id": {
											"type": "number",
											"format": "double"
										}
									},
									"required": [
										"updatedAt",
										"createdAt",
										"isActive",
										"role",
										"email",
										"password",
										"username",
										"id"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"updatedAt": {
											"type": "string",
											"format": "date-time"
										},
										"createdAt": {
											"type": "string",
											"format": "date-time"
										},
										"isActive": {
											"type": "boolean"
										},
										"role": {
											"$ref": "#/components/schemas/_36_Enums.UserRole"
										},
										"email": {
											"type": "string"
										},
										"password": {
											"type": "string"
										},
										"username": {
											"type": "string"
										},
										"id": {
											"type": "number",
											"format": "double"
										}
									},
									"required": [
										"updatedAt",
										"createdAt",
										"isActive",
										"role",
										"email",
										"password",
										"username",
										"id"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateUserDTO"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"updatedAt": {
											"type": "string",
											"format": "date-time"
										},
										"createdAt": {
											"type": "string",
											"format": "date-time"
										},
										"isActive": {
											"type": "boolean"
										},
										"role": {
											"$ref": "#/components/schemas/_36_Enums.UserRole"
										},
										"email": {
											"type": "string"
										},
										"password": {
											"type": "string"
										},
										"username": {
											"type": "string"
										},
										"id": {
											"type": "number",
											"format": "double"
										}
									},
									"required": [
										"updatedAt",
										"createdAt",
										"isActive",
										"role",
										"email",
										"password",
										"username",
										"id"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/departments": {
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Department"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Lấy danh sách toàn bộ khoa",
				"tags": [
					"Department"
				],
				"security": [],
				"parameters": []
			}
		},
		"/departments/{id}": {
			"get": {
				"operationId": "GetById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Department"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"description": "Lấy chi tiết một khoa theo ID",
				"tags": [
					"Department"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Department"
								}
							}
						}
					}
				},
				"description": "Cập nhật thông tin khoa",
				"tags": [
					"Department"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DepartmentUpdateRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"204": {
						"description": "No Content"
					}
				},
				"description": "Xóa một khoa khỏi hệ thống",
				"tags": [
					"Department"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/departments/create": {
			"post": {
				"operationId": "Create",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Department"
								}
							}
						}
					}
				},
				"description": "Tạo mới một khoa",
				"tags": [
					"Department"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DepartmentCreateRequest"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/api"
		}
	]
}